Add comments and missing option to template
When working on a system under presure its good to be able to have some context
to the options in the file. In this specific case the options are fairly self
explanitory but having the comment helps explain what they are meant to do.

This change adds the comments with the relevant option so ony those related to
(eg) git show when (eg) etckeeper_git_commit_options is set.

This also adds an option for etckeeper_avoid_special_file_warning which could
be a separate branch but seemed unintrusive enough to bundle in with this set
of changes. Its only relevant for users running the command manually (as I am
at the moment) so shouldn't affect the role very much.